convert zset button to poi planarize "P" button
this sets poi1 without zeroing the system.

We leave poi1 as zeroing because normally this is a good thing -
this keeps our chip images aligned to a 0 coordinate system.

only planarization wants an absolute zero, so an extra button
is created to set POI 1 that also doesn't zero the coordinate system